refactor: layout에서 common Header, Footer 사용 (#11)
