+++ src/components/common/Dropdown.tsx
@@ -1,5 +1,6 @@
 'use client'
 
+import React from 'react'
 import Link from 'next/link'
 
 // Dropdown 공통 컴포넌트
